Removed checking for passwd and will update README
The password checking functionality at the beginning did not work as expected for a new user without a password set, so I removed it and updated the README to advise either setting the user password through the KDE Plasma GUI, or manually running passwd from a command line, if the user knows that a username password has not been set yet on his or her Steam Deck. No other functional change was made.