Updated Installation script
The password checking for a new user now works and prompts to set a password for the deck user if one has not already been set. The script now also checks for a duplicate rEFInd EFI boot entry from previous script runs and deletes it (if found). The script can now be run back to back without any intermediate steps. If reinstalling Windows, still disable the rEFInd EFI entry beforehand (as described in the README).