Added lines for background options
Uncomment the line for the desired background (and comment the others out). PNG format seems to be most friendly and compatible for icons and backgrounds. If you have a 1,280x800 JPG file that you want to use as a background, I recommend using GIMP (or other photo editing software) to export it as a PNG file.