No Pacman installation script improvements
I've added a check for missing SteamOS EFI entry which happens with some BIOS updates. The script will check if it's missing or not, and if it's missing will recreate a new "SteamOS" EFI entry.  This also adds the systemctl daemon that makes rEFInd the next boot top priority every time SteamOS boots. This is part of a workaround (or fix even) that helps the rEFInd config survive SteamOS branch changes and SteamOS updates (excluding BIOS updates).